ofs_new/telegram_bot/requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
pydantic>=1.9.0 